+++ ali/CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added support for Forge's `BasicItemListing` trades
 - Added support for Forge's ingredients (compound, difference, intersection, partial NBT, strict NBT)
 - Added support for Fabric's difference and NBT ingredients
+- Fixed recipe screenshot not working on EMI
 
 ## [1.12.0]